Sunday (21.04.2013)
Features added:
-Sqlite database on client side, all treasures are stored in a table,
also there is additional table for storing media information (photos)
encoded as a string.
- MapActivity - represents treasures on a map according to their
coordinates and current location of the device. It can show all
treasures at a time, or a specific treasure ( in case of redirecting
from ShowTreasureActivity). Each treasure is represented by marker. If u
want to see short description of a treasure, just click on it and a
small InfoWindow will be shown! To get more detailed information about
the treasure, user has to click on a InfoWindow what will redirects him
to ShowTreasureActivity.
- Improved navigation between activities